+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_api_container_functions.go
@@ -5,6 +5,7 @@ import (
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_operation_parallelizer"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/port_spec_serializer"
@@ -19,9 +20,14 @@ import (
 )
 
 const (
+	// TODO Remove this when we switch fully to the data volume
 	// The location where the enclave data directory (on the Docker host machine) will be bind-mounted
 	//  on the API container
-	enclaveDataDirpathOnAPIContainer = "/kurtosis-enclave-data"
+	enclaveDataBindmountDirpathOnAPIContainer = "/kurtosis-enclave-data"
+
+	// The location where the enclave data volume will be mounted
+	//  on the API container
+	enclaveDataVolumeDirpathOnAPIContainer = "/kurtosis-data"
 
 	// The API container uses gRPC so MUST listen on TCP (no other protocols are supported), which also
 	// means that its grpc-proxy must listen on TCP
@@ -44,6 +50,7 @@ func (backend *DockerKurtosisBackend) CreateAPIContainer(
 	ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
 	grpcPortNum uint16,
 	grpcProxyPortNum uint16,
+	// TODO remove when we switch fully to enclave data volume
 	enclaveDataDirpathOnHostMachine string,
 	envVars map[string]string,
 ) (*api_container.APIContainer, error) {
@@ -59,6 +66,11 @@ func (backend *DockerKurtosisBackend) CreateAPIContainer(
 	}
 	if len(preexistingApiContainersInEnclave) > 0 {
 		return nil, stacktrace.NewError("Found existing API container(s) in enclave '%v'; cannot start a new one", enclaveId)
+	}
+
+	enclaveDataVolumeName, err := backend.getEnclaveDataVolumeByEnclaveId(ctx, enclaveId)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the enclave data volume for enclave '%v'", enclaveId)
 	}
 
 	// Get the Docker network ID where we'll start the new API container
@@ -118,7 +130,11 @@ func (backend *DockerKurtosisBackend) CreateAPIContainer(
 	bindMounts := map[string]string{
 		// Necessary so that the API container can interact with the Docker engine
 		dockerSocketFilepath:            dockerSocketFilepath,
-		enclaveDataDirpathOnHostMachine: enclaveDataDirpathOnAPIContainer,
+		enclaveDataDirpathOnHostMachine: enclaveDataBindmountDirpathOnAPIContainer,
+	}
+
+	volumeMounts := map[string]string{
+		enclaveDataVolumeName: enclaveDataVolumeDirpathOnAPIContainer,
 	}
 
 	labelStrs := map[string]string{}
@@ -134,6 +150,8 @@ func (backend *DockerKurtosisBackend) CreateAPIContainer(
 		envVars,
 	).WithBindMounts(
 		bindMounts,
+	).WithVolumeMounts(
+		volumeMounts,
 	).WithUsedPorts(
 		usedPorts,
 	).WithStaticIP(
@@ -226,30 +244,41 @@ func (backend *DockerKurtosisBackend) StopAPIContainers(
 		return nil, nil, stacktrace.Propagate(err, "An error occurred getting API containers matching filters '%+v'", filters)
 	}
 
-	containerIdsToKill := map[string]bool{}
-	for containerId := range matchingApiContainersByContainerId {
-		containerIdsToKill[containerId] = true
+	// TODO PLEAAASE GO GENERICS... but we can't use 1.18 yet because it'll break all Kurtosis clients :(
+	matchingUncastedApiContainersByContainerId := map[string]interface{}{}
+	for containerId, apiContainerObj := range matchingApiContainersByContainerId {
+		matchingUncastedApiContainersByContainerId[containerId] = interface{}(apiContainerObj)
 	}
 
-	successfulContainerIds, erroredContainerIds := backend.killContainers(ctx, containerIdsToKill)
+	var killApiContainerOperation docker_operation_parallelizer.DockerOperation = func(
+		ctx context.Context,
+		dockerManager *docker_manager.DockerManager,
+		dockerObjectId string,
+	) error {
+		if err := dockerManager.KillContainer(ctx, dockerObjectId); err != nil {
+			return stacktrace.Propagate(err, "An error occurred killing API container with ID '%v'", dockerObjectId)
+		}
+		return nil
+	}
+
+	successfulEnclaveIdStrs, erroredEnclaveIdStrs, err := docker_operation_parallelizer.RunDockerOperationInParallelForKurtosisObjects(
+		ctx,
+		matchingUncastedApiContainersByContainerId,
+		backend.dockerManager,
+		extractEnclaveIdFromUncastedApiContainerObj,
+		killApiContainerOperation,
+	)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred killing API containers matching filters '%+v'", filters)
+	}
 
 	successfulEnclaveIds := map[enclave.EnclaveID]bool{}
-	for containerId := range successfulContainerIds {
-		apiContainerObj, found := matchingApiContainersByContainerId[containerId]
-		if !found {
-			return nil, nil, stacktrace.NewError("Successfully killed container with ID '%v' that wasn't requested; this is a bug in Kurtosis!", containerId)
-		}
-		successfulEnclaveIds[apiContainerObj.GetEnclaveID()] = true
+	for enclaveIdStr := range successfulEnclaveIdStrs {
+		successfulEnclaveIds[enclave.EnclaveID(enclaveIdStr)] = true
 	}
-
 	erroredEnclaveIds := map[enclave.EnclaveID]error{}
-	for containerId := range erroredContainerIds {
-		apiContainerObj, found := matchingApiContainersByContainerId[containerId]
-		if !found {
-			return nil, nil, stacktrace.NewError("An error occurred killing container with ID '%v' that wasn't requested; this is a bug in Kurtosis!", containerId)
-		}
-		wrappedErr := stacktrace.Propagate(err, "An error occurred killing API container in enclave '%v' with container ID '%v'", apiContainerObj.GetEnclaveID(), containerId)
-		erroredEnclaveIds[apiContainerObj.GetEnclaveID()] = wrappedErr
+	for enclaveIdStr, killErr := range erroredEnclaveIdStrs {
+		erroredEnclaveIds[enclave.EnclaveID(enclaveIdStr)] = killErr
 	}
 
 	return successfulEnclaveIds, erroredEnclaveIds, nil
@@ -261,23 +290,44 @@ func (backend *DockerKurtosisBackend) DestroyAPIContainers(ctx context.Context, 
 		return nil, nil, stacktrace.Propagate(err, "An error occurred getting API containers matching the following filters: %+v", filters)
 	}
 
-	successIds := map[enclave.EnclaveID]bool{}
-	errorIds := map[enclave.EnclaveID]error{}
-	for containerId, engineObj := range matchingApiContainersByContainerId {
-		enclaveId := engineObj.GetEnclaveID()
-		if err := backend.dockerManager.RemoveContainer(ctx, containerId); err != nil {
-			wrappedErr := stacktrace.Propagate(
-				err,
-				"An error occurred removing API container for enclave '%v' with container ID '%v'",
-				enclaveId,
-				containerId,
-			)
-			errorIds[enclaveId] = wrappedErr
-		} else {
-			successIds[enclaveId] = true
-		}
+	// TODO PLEAAASE GO GENERICS... but we can't use 1.18 yet because it'll break all Kurtosis clients :(
+	matchingUncastedApiContainersByContainerId := map[string]interface{}{}
+	for containerId, apiContainerObj := range matchingApiContainersByContainerId {
+		matchingUncastedApiContainersByContainerId[containerId] = interface{}(apiContainerObj)
 	}
-	return successIds, errorIds, nil
+
+	var removeApiContainerOperation docker_operation_parallelizer.DockerOperation = func(
+		ctx context.Context,
+		dockerManager *docker_manager.DockerManager,
+		dockerObjectId string,
+	) error {
+		if err := dockerManager.RemoveContainer(ctx, dockerObjectId); err != nil {
+			return stacktrace.Propagate(err, "An error occurred removing API container with ID '%v'", dockerObjectId)
+		}
+		return nil
+	}
+
+	successfulEnclaveIdStrs, erroredEnclaveIdStrs, err := docker_operation_parallelizer.RunDockerOperationInParallelForKurtosisObjects(
+		ctx,
+		matchingUncastedApiContainersByContainerId,
+		backend.dockerManager,
+		extractEnclaveIdFromUncastedApiContainerObj,
+		removeApiContainerOperation,
+	)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred removing API containers matching filters '%+v'", filters)
+	}
+
+	successfulEnclaveIds := map[enclave.EnclaveID]bool{}
+	for enclaveIdStr := range successfulEnclaveIdStrs {
+		successfulEnclaveIds[enclave.EnclaveID(enclaveIdStr)] = true
+	}
+	erroredEnclaveIds := map[enclave.EnclaveID]error{}
+	for enclaveIdStr, killErr := range erroredEnclaveIdStrs {
+		erroredEnclaveIds[enclave.EnclaveID(enclaveIdStr)] = killErr
+	}
+
+	return successfulEnclaveIds, erroredEnclaveIds, nil
 }
 
 // ====================================================================================================
@@ -441,4 +491,12 @@ func getPrivateApiContainerPorts(containerLabels map[string]string) (
 	}
 
 	return grpcPortSpec, grpcProxyPortSpec, nil
+}
+
+func extractEnclaveIdFromUncastedApiContainerObj(uncastedApiContainerObj interface{}) (string, error) {
+	castedObj, ok := uncastedApiContainerObj.(*api_container.APIContainer)
+	if !ok {
+		return "", stacktrace.NewError("An error occurred downcasting the API container object")
+	}
+	return string(castedObj.GetEnclaveID()), nil
 }