+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,24 @@
 # TBD
 
+# 0.15.1
+### Features
+* The enclave data volume gets mounted on all services
+* `DockerKurtosisBackend.CreateEnclave` also creates an enclave data volume
+* Parallelized several operations to improve perf:
+    * `DockerKurtosisBackend.StopEnclaves`
+    * `DockerKurtosisBackend.DestroyEnclaves`
+    * `DockerKurtosisBackend.StopAPIContainers`
+    * `DockerKurtosisBackend.DestroyAPIContainers`
+    * `DockerKurtosisBackend.StopEngines`
+    * `DockerKurtosisBackend.DestroyEngines`
+    * `DockerKurtosisBackend.StopModules`
+    * `DockerKurtosisBackend.DestroyModules`
+    * `DockerKurtosisBackend.StopNetworkingSidecars`
+    * `DockerKurtosisBackend.DestroyNetworkingSidecars`
+    * `DockerKurtosisBackend.StopUserServices`
+    * `DockerKurtosisBackend.DestroyUserServices`
+
+
 # 0.15.0
 ### Fixes
 * Fixed `UserService` object not having a `GetPrivatePorts` method